challenge-2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ria"
version = "0.1.0"
description = "Reactive Investment Agent — challenge-2 dogfood CLI"
requires-python = ">=3.11"
authors = [{ name = "biweekly-ai" }]
readme = { text = "Reactive Investment Agent (RIA) — see challenge-2/PRD.md", content-type = "text/plain" }
license = { text = "UNLICENSED" }
dependencies = [
    "anthropic~=0.96.0",
    "typer~=0.24.0",
    "sentence-transformers~=5.4.0",
    "psycopg[binary]~=3.3.0",
    "pgvector~=0.4.0",
    "pandas~=3.0.0",
    "pandas_market_calendars~=5.3.0",
    "pyyaml~=6.0",
    "pydantic>=2.13,<3",
    "pytest~=9.0.0",
    "pytest-vcr~=1.0.0",
    "yfinance~=1.3.0",
    "requests~=2.33.0",
]

[project.scripts]
ria = "ria.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
